% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mod2rm.r
\name{summary.mod2rm}
\alias{summary.mod2rm}
\title{Print and summary function for objects of class "mod2rm"}
\usage{
\method{summary}{mod2rm}(object, ...)
}
\arguments{
\item{object}{An object of class "mod2rm"}

\item{...}{Additional parameters}
}
\value{
Prints summary of the object, then returns NULL
}
\description{
Prints a summary of a list object of class "mod2rm"
}
\examples{
df = data.frame(out1 = c(2,4,5,6,6), out2 = c(7,4,5,1,1), w = c(9,4,4,2,3))
res = mod2rm(df, out1, out2, w)
res = mod2rm(df, out1, out2, w, c(2,5))
summary.mod2rm(res)
}
